refactor: viewModel 생성시 instance container를 주입할 수 있는 형식으로 변경한다.
