+++ app/src/main/java/woowacourse/shopping/ShoppingApplication.kt
@@ -1,9 +1,9 @@
 package woowacourse.shopping
 
 import android.app.Application
-import com.boogiwoogi.di.version2.DefaultDiContainer
-import com.boogiwoogi.di.version2.DiContainer
+import com.boogiwoogi.di.version2.DefaultInstanceContainer
 import com.boogiwoogi.di.version2.DiInjector
+import com.boogiwoogi.di.version2.InstanceContainer
 import com.boogiwoogi.di.version2.Modules
 import woowacourse.shopping.di.application.ApplicationModule
 
@@ -12,15 +12,15 @@ class ShoppingApplication : Application() {
     override fun onCreate() {
         super.onCreate()
 
-        container = DefaultDiContainer()
         modules = ApplicationModule(this)
+        container = DefaultInstanceContainer()
         injector = DiInjector()
     }
 
     companion object {
 
-        lateinit var container: DiContainer
         lateinit var modules: Modules
+        lateinit var container: InstanceContainer
         lateinit var injector: DiInjector
     }
 }
