+++ app/src/main/java/woowacourse/shopping/di/viewmodel/DiViewModelInitializer.kt
@@ -1,4 +1,4 @@
-package com.boogiwoogi.di.version2
+package woowacourse.shopping.di.viewmodel
 
 import androidx.activity.ComponentActivity
 import androidx.annotation.MainThread
@@ -7,18 +7,21 @@ import androidx.lifecycle.ViewModelLazy
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewmodel.initializer
 import androidx.lifecycle.viewmodel.viewModelFactory
+import com.boogiwoogi.di.version2.InstanceContainer
 import woowacourse.shopping.ShoppingApplication
 
 @MainThread
-inline fun <reified VM : ViewModel> ComponentActivity.woogiViewModels(): Lazy<VM> {
+inline fun <reified VM : ViewModel> ComponentActivity.woogiViewModels(
+    container: InstanceContainer = ShoppingApplication.container
+): Lazy<VM> {
     return ViewModelLazy(
         VM::class,
         { viewModelStore },
-        { woogiViewModelFactory<VM>() },
+        { woogiViewModelFactory<VM>(container) },
     )
 }
 
-inline fun <reified VM : ViewModel> woogiViewModelFactory(): ViewModelProvider.Factory =
+inline fun <reified VM : ViewModel> woogiViewModelFactory(container: InstanceContainer): ViewModelProvider.Factory =
     viewModelFactory {
         initializer {
             ShoppingApplication.run {
